did some fixes
