+++ BankexWallet/InterfaceElements/DesignElements/WalletSettings/Wallets/View/WalletCell.swift
